Making the repo neat
All the individual files in this folder was added initially the person named beside the file and the final_assembler was added by somay jalan. Now somay jalan removed all the files to make the repo neat and make space for simulator files. If you want to check records you can go in the commit history and check that these were initially committed by these individuals.